requirements.txt
pyserial==3.3
rtmidi-python==0.2.2
